requirements.txt
appdirs==1.4.3
beautifulsoup4==4.5.3
Django==1.10.6
django-crispy-forms==1.6.1
django-extensions==1.7.8
django-filter==1.0.2
django-livereload-server==0.2.3
django-registration-redux==1.6
djangorestframework==3.6.2
livereload==2.5.1
Markdown==2.6.8
packaging==16.8
pydot==1.2.3
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
tornado==4.4.3
